# docs/knowledge/backend/swagger.yaml
openapi: 3.0.0
paths:
...
  /api/visitors/check-by-phone:
    get:
      operationId: VisitorsController_findVisitorByPhone
      parameters:
        - name: phone
          required: true
          in: query
          description: Visitor's 10-digit phone number
          schema:
            type: string
        - name: branchId
          required: true
          in: query
          description: The ID of the branch being visited
          schema:
            type: number
      responses:
        "200":
          description: ""
      summary: Check if a visitor exists by phone number for a specific branch
      tags: &a7
        - Visitor
  /api/visitors/register:
    post:
      operationId: VisitorsController_registerVisitor
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/RegisterVisitorDto"
      responses:
        "201":
          description: ""
      security:
        - access-token: []
      summary: Register a new visitor with optional documents
      tags: *a7
  /api/visitors/request-visit:
    post:
      operationId: VisitorsController_createVisitRequest
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateVisitRequestDto"
      responses:
        "201":
          description: ""
      security:
        - access-token: []
      summary: Create a visit request for a registered visitor (by Security)
      tags: *a7
  /api/visitors/verify-code:
    post:
      operationId: VisitorsController_verifyCode
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyVisitCodeDto"
      responses:
        "201":
          description: ""
      security:
        - access-token: []
      summary: Check-in visitor using visit code (by Security)
      tags: *a7
  /api/visitors/verify-checkin-otp:
    post:
      operationId: VisitorsController_verifyCheckInOtp
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyCheckInOtpDto"
      responses:
        "200":
          description: OTP verified successfully
        "400":
          description: Invalid OTP or expired
        "401":
          description: Unauthorized
        "404":
          description: Visit not found
      security:
        - access-token: []
      summary: Verify Check-In OTP for visitor at gate (Security only)
      tags: *a7
  /api/visitors/checkin/{visitId}:
    post:
      operationId: VisitorsController_checkInVisitor
      parameters:
        - name: visitId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Visitor checked in successfully
        "400":
          description: Invalid visit state for check-in
        "401":
          description: Unauthorized
        "403":
          description: Forbidden - Visit not in your branch
        "404":
          description: Visit not found
      security:
        - access-token: []
      summary: Check-in visitor using visit ID (after OTP verification)
      tags: *a7
  /api/visitors/search:
    get:
      operationId: VisitorsController_searchVisitors
      parameters:
        - name: phone
          required: true
          in: query
          description: 10-digit phone number to search for
          schema:
            minLength: 10
            maxLength: 10
            example: "9876543210"
            type: string
        - name: branchId
          required: true
          in: query
          description: Branch ID to filter visitors
          schema:
            example: branch-uuid-123
            type: string
      responses:
        "200":
          description: Search completed successfully
        "400":
          description: Invalid request parameters
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
      security:
        - access-token: []
      summary: Search for a visitor by phone number (Security only)
      tags: *a7
  /api/visitors/scan-qr:
    post:
      operationId: VisitorsController_scanQRCode
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyVisitCodeDto"
      responses:
        "201":
          description: ""
      summary: Check-in visitor by scanning QR code (by Security)
      tags: *a7
  /api/visitors/checkout/{visitId}:
    patch:
      operationId: VisitorsController_checkout
      parameters:
        - name: visitId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security:
        - access-token: []
      summary: Check-out visitor (by Security)
      tags: *a7
  /api/visitors/active:
    get:
      operationId: VisitorsController_listActive
      parameters: []
      responses:
        "200":
          description: ""
      security:
        - access-token: []
      summary: List active (checked-in) visitors (for Security)
      tags: *a7
  /api/visitors/summary:
    get:
      operationId: VisitorsController_summary
      parameters:
        - name: date
          required: false
          in: query
          description: Filter by a single date (YYYY-MM-DD)
          schema:
            type: string
        - name: startDate
          required: false
          in: query
          description: Start date for a date range (YYYY-MM-DD)
          schema:
            type: string
        - name: endDate
          required: false
          in: query
          description: End date for a date range (YYYY-MM-DD)
          schema:
            type: string
        - name: status
          required: false
          in: query
          description: Filter by visit status
          schema:
            type: string
        - name: personToMeet
          required: false
          in: query
          description: Filter by the staff member the visitor met
          schema:
            type: string
        - name: search
          required: false
          in: query
          description: Search by visitor name or phone number
          schema:
            type: string
        - name: page
          required: false
          in: query
          description: Page number for pagination
          schema:
            type: number
        - name: limit
          required: false
          in: query
          description: Number of items per page
          schema:
            type: number
        - name: skip
          required: false
          in: query
          description: Number of items to skip for pagination
          schema:
            type: number
        - name: take
          required: false
          in: query
          description: Number of items to take for pagination
          schema:
            type: number
      responses:
        "200":
          description: ""
      security:
        - access-token: []
      summary: Get visitor summary for the branch
      tags: *a7
  /api/visitors/update/{visitorId}:
    patch:
      operationId: VisitorsController_updateVisitor
      parameters:
        - name: visitorId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/RegisterVisitorDto"
      responses:
        "200":
          description: ""
      security:
        - access-token: []
      summary: Update visitor details and documents
      tags: *a7
  /api/visitors/download/{visitorId}/{documentType}:
    get:
      operationId: VisitorsController_downloadDocument
      parameters:
        - name: visitorId
          required: true
          in: path
          schema:
            type: string
        - name: documentType
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: File downloaded successfully
        "403":
          description: Access denied
        "404":
          description: Visitor or document not found
      security:
        - access-token: []
      summary: Download visitor document
      tags: *a7
  /api/public/visitors/check-by-phone:
    get:
      operationId: PublicVisitorsController_findVisitorByPhone
      parameters:
        - name: phone
          required: true
          in: query
          description: Visitor's 10-digit phone number
          schema:
            type: string
        - name: branchId
          required: true
          in: query
          description: Branch ID from QR code
          schema:
            type: string
      responses:
        "200":
          description: ""
      summary: Check if a visitor exists by phone number for a specific branch
      tags: &a8
        - Public Visitor
...
  /api/public/visitors/register:
    post:
      operationId: PublicVisitorsController_registerLegacyVisitor
      parameters:
        - name: branchId
          required: true
          in: query
          description: Branch ID from QR code
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/PublicRegisterVisitorDto"
      responses:
        "201":
          description: ""
      summary: Register a new visitor via public form (full registration for meeting
        visitors)
      tags: *a8
...
  /api/public/visitors/request-visit:
    post:
      operationId: PublicVisitorsController_createVisitRequest
      parameters:
        - name: branchId
          required: true
          in: query
          description: Branch ID from QR code
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/PublicCreateVisitRequestDto"
      responses:
        "201":
          description: ""
      summary: Create a visit request via public form (legacy endpoint)
      tags: *a8
  /api/public/visitors/update/{visitorId}:
    patch:
      operationId: PublicVisitorsController_updateVisitor
      parameters:
        - name: visitorId
          required: true
          in: path
          schema:
            type: string
        - name: branchId
          required: true
          in: query
          description: Branch ID from QR code
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/PublicRegisterVisitorDto"
      responses:
        "200":
          description: ""
      summary: Update visitor details via public form (from QR code)
      tags: *a8
...
info:
  title: Hospital Visitor Tracking System
  description: API documentation for the Hospital Visitor Tracking System
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    RegisterVisitorDto:
      type: object
      properties:
        phone:
          type: string
        branchId:
          type: string
        firstName:
          type: string
        middleName:
          type: string
        lastName:
          type: string
        alternatePhone:
          type: string
        email:
          type: string
        alternateEmail:
          type: string
        address:
          type: string
        company:
          type: string
        companyWebsite:
          type: string
        designation:
          type: string
        photo:
          type: string
          format: binary
          description: Visitor photo (.jpg, .png)
        governmentIdDocument:
          type: string
          format: binary
          description: Government ID Document (.jpg, .png, .pdf)
        officeIdDocument:
          type: string
          format: binary
          description: Office ID Document (.jpg, .png, .pdf)
        reportingManagerName:
          type: string
          description: Reporting Manager name
        reportingManagerPhone:
          type: string
          description: Reporting Manager phone
      required:
        - phone
        - branchId
        - firstName
        - lastName
    CreateVisitRequestDto:
      type: object
      properties:
        phone:
          type: string
          description: Visitor's registered phone number
        purpose:
          type: string
        department:
          type: string
          enum:
            - GENERAL_MEDICINE
            - CARDIOLOGY
            - NEUROLOGY
            - PULMONOLOGY
            - NEPHROLOGY
            - GASTROENTEROLOGY
            - ENDOCRINOLOGY
            - ONCOLOGY
            - HEMATOLOGY
            - DERMATOLOGY
            - INFECTIOUS_DISEASE
            - RHEUMATOLOGY
            - OPHTHALMOLOGY
            - DENTISTRY
            - PAIN_MANAGEMENT
            - GENERAL_SURGERY
            - ORTHOPEDICS
            - NEUROSURGERY
            - CARDIOTHORACIC_SURGERY
            - ENT_SURGERY
            - PLASTIC_RECONSTRUCTIVE_SURGERY
            - UROLOGY
            - OBSTETRICS_GYNECOLOGY
            - PEDIATRICS
            - NEONATOLOGY
            - EMERGENCY
            - INTENSIVE_CARE_UNIT
            - CRITICAL_CARE_UNIT
            - TRAUMA_CENTER
            - RADIOLOGY
            - PATHOLOGY
            - PHARMACY
            - BLOOD_BANK
            - ANESTHESIOLOGY
            - RESPIRATORY_CARE
            - CENTRAL_SUPPLY
            - NUTRITION_AND_DIETETICS
            - MEDICAL_RECORDS
            - SECURITY
            - ADMINISTRATION
            - HUMAN_RESOURCES
            - FINANCE
            - IT_SUPPORT
            - LEGAL_COMPLIANCE
        personToMeet:
          type: string
          description: ID of the staff member to meet, or "other" for unlisted staff.
            Required if staffName and staffPhone are not provided.
        staffName:
          type: string
          description: Name of the staff to meet. Required if personToMeet is "other" or
            not provided.
        staffPhone:
          type: string
          description: Phone of the staff to meet. Required if personToMeet is "other" or
            not provided.
        visitingCardPhoto:
          type: string
          description: Base64 encoded visiting card photo
      required:
        - phone
        - purpose
    VerifyVisitCodeDto:
      type: object
      properties:
        phone:
          type: string
        visitCode:
          type: string
      required:
        - phone
        - visitCode
    VerifyCheckInOtpDto:
      type: object
      properties:
        otp:
          type: string
          description: 6-digit Check-In OTP provided by visitor
          example: "654321"
          minLength: 6
          maxLength: 6
        branchId:
          type: string
          description: Branch ID where the visit is taking place
          example: branch-uuid-123
      required:
        - otp
        - branchId
    PublicRegisterVisitorDto:
      type: object
      properties:
        phone:
          type: string
        firstName:
          type: string
        middleName:
          type: string
        lastName:
          type: string
        alternatePhone:
          type: string
        email:
          type: string
        alternateEmail:
          type: string
        address:
          type: string
        company:
          type: string
        companyWebsite:
          type: string
        designation:
          type: string
        reportingManagerName:
          type: string
          description: Reporting Manager name
        reportingManagerPhone:
          type: string
          description: Reporting Manager phone
        photo:
          type: string
          format: binary
          description: Visitor photo (.jpg, .png)
        governmentIdDocument:
          type: string
          format: binary
          description: Government ID Document (.jpg, .png, .pdf)
        officeIdDocument:
          type: string
          format: binary
          description: Office ID Document (.jpg, .png, .pdf)
      required:
        - phone
        - firstName
        - lastName
...
    PublicCreateVisitRequestDto:
      type: object
      properties:
        phone:
          type: string
          description: Visitor's registered phone number
        purpose:
          type: string
        department:
          type: string
          enum:
            - GENERAL_MEDICINE
            - CARDIOLOGY
            - NEUROLOGY
            - PULMONOLOGY
            - NEPHROLOGY
            - GASTROENTEROLOGY
            - ENDOCRINOLOGY
            - ONCOLOGY
            - HEMATOLOGY
            - DERMATOLOGY
            - INFECTIOUS_DISEASE
            - RHEUMATOLOGY
            - OPHTHALMOLOGY
            - DENTISTRY
            - PAIN_MANAGEMENT
            - GENERAL_SURGERY
            - ORTHOPEDICS
            - NEUROSURGERY
            - CARDIOTHORACIC_SURGERY
            - ENT_SURGERY
            - PLASTIC_RECONSTRUCTIVE_SURGERY
            - UROLOGY
            - OBSTETRICS_GYNECOLOGY
            - PEDIATRICS
            - NEONATOLOGY
            - EMERGENCY
            - INTENSIVE_CARE_UNIT
            - CRITICAL_CARE_UNIT
            - TRAUMA_CENTER
            - RADIOLOGY
            - PATHOLOGY
            - PHARMACY
            - BLOOD_BANK
            - ANESTHESIOLOGY
            - RESPIRATORY_CARE
            - CENTRAL_SUPPLY
            - NUTRITION_AND_DIETETICS
            - MEDICAL_RECORDS
            - SECURITY
            - ADMINISTRATION
            - HUMAN_RESOURCES
            - FINANCE
            - IT_SUPPORT
            - LEGAL_COMPLIANCE
        personToMeet:
          type: string
          description: ID of the staff member to meet, or "other" for unlisted staff.
            Required if staffName and staffPhone are not provided.
        staffName:
          type: string
          description: Name of the staff to meet. Required if personToMeet is "other" or
            not provided.
        staffPhone:
          type: string
          description: Phone of the staff to meet. Required if personToMeet is "other" or
            not provided.
        visitingCardPhoto:
          type: string
          description: Base64 encoded visiting card photo
      required:
        - phone
        - purpose
...